# docs/openapi.json
{
  "components": {
    "schemas": {
      "BatchPredictItem": {
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "response": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PredictResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "ok"
        ],
        "title": "BatchPredictItem",
        "type": "object"
      },
      "BatchPredictRequest": {
        "properties": {
          "items": {
            "description": "One PredictRequest per user. Capped at 2000 per call.",
            "items": {
              "$ref": "#/components/schemas/PredictRequest"
            },
            "maxItems": 2000,
            "minItems": 1,
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "BatchPredictRequest",
        "type": "object"
      },
      "BatchPredictResponse": {
        "properties": {
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "n_failed": {
            "title": "N Failed",
            "type": "integer"
          },
          "n_ok": {
            "title": "N Ok",
            "type": "integer"
          },
          "n_users": {
            "title": "N Users",
            "type": "integer"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/BatchPredictItem"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "model_version",
          "n_users",
          "n_ok",
          "n_failed",
          "results"
        ],
        "title": "BatchPredictResponse",
        "type": "object"
      },
...
      "DoseEvent": {
        "additionalProperties": false,
        "properties": {
          "dose_class": {
            "default": "other",
            "enum": [
              "cardio",
              "neuro",
              "endocrine",
              "psych",
              "antibiotic",
              "supplement",
              "other"
            ],
            "title": "Dose Class",
            "type": "string"
          },
          "dose_id": {
            "title": "Dose Id",
            "type": "string"
          },
          "dose_strength_mg": {
            "default": 0.0,
            "title": "Dose Strength Mg",
            "type": "number"
          },
          "scheduled_at": {
            "format": "date-time",
            "title": "Scheduled At",
            "type": "string"
          },
          "status": {
            "enum": [
              "taken",
              "missed",
              "skipped",
              "late"
            ],
            "title": "Status",
            "type": "string"
          },
          "taken_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taken At"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "dose_id",
          "scheduled_at",
          "status"
        ],
        "title": "DoseEvent",
        "type": "object"
      },
      "DosePrediction": {
        "properties": {
          "dose_id": {
            "title": "Dose Id",
            "type": "string"
          },
          "miss_probability": {
            "title": "Miss Probability",
            "type": "number"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/ReasonCode"
            },
            "title": "Reasons",
            "type": "array"
          },
          "risk_tier": {
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Risk Tier",
            "type": "string"
          },
          "scheduled_at": {
            "format": "date-time",
            "title": "Scheduled At",
            "type": "string"
          }
        },
        "required": [
          "dose_id",
          "scheduled_at",
          "miss_probability",
          "risk_tier"
        ],
        "title": "DosePrediction",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PredictRequest": {
        "properties": {
          "history": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DoseEvent"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "History"
          },
          "schedule": {
            "items": {
              "$ref": "#/components/schemas/ScheduledDose"
            },
            "title": "Schedule",
            "type": "array"
          },
          "top_k_reasons": {
            "default": 3,
            "title": "Top K Reasons",
            "type": "integer"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id"
        ],
        "title": "PredictRequest",
        "type": "object"
      },
      "PredictResponse": {
        "properties": {
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "predictions": {
            "items": {
              "$ref": "#/components/schemas/DosePrediction"
            },
            "title": "Predictions",
            "type": "array"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "model_version",
          "predictions"
        ],
        "title": "PredictResponse",
        "type": "object"
      },
      "ReasonCode": {
        "properties": {
          "contribution": {
            "title": "Contribution",
            "type": "number"
          },
          "feature": {
            "title": "Feature",
            "type": "string"
          },
          "human": {
            "title": "Human",
            "type": "string"
          }
        },
        "required": [
          "feature",
          "contribution",
          "human"
        ],
        "title": "ReasonCode",
        "type": "object"
      },
      "ScheduledDose": {
        "additionalProperties": false,
        "properties": {
          "dose_class": {
            "default": "other",
            "enum": [
              "cardio",
              "neuro",
              "endocrine",
              "psych",
              "antibiotic",
              "supplement",
              "other"
            ],
            "title": "Dose Class",
            "type": "string"
          },
          "dose_id": {
            "title": "Dose Id",
            "type": "string"
          },
          "dose_strength_mg": {
            "default": 0.0,
            "title": "Dose Strength Mg",
            "type": "number"
          },
          "scheduled_at": {
            "format": "date-time",
            "title": "Scheduled At",
            "type": "string"
          }
        },
        "required": [
          "dose_id",
          "scheduled_at"
        ],
        "title": "ScheduledDose",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Predicts which medication doses a user is likely to miss in the next 24 hours. Backed by XGBoost + LightGBM ensemble with SHAP explanations.",
    "title": "adherence-ml",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/predict": {
      "post": {
        "operationId": "predict_v1_predict_post",
        "parameters": [
          {
            "in": "query",
            "name": "model_name",
            "required": false,
            "schema": {
              "default": "default",
              "title": "Model Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Predict",
        "tags": [
          "predict"
        ]
      }
    },
    "/v1/predict/batch": {
      "post": {
        "description": "Score upcoming doses for many users in one call.\n\nDesigned for Med-Tracker's nightly cron: send {items: [PredictRequest, ...]}\nand get back per-user predictions + reason codes. Errors are isolated to\nthe offending user (the batch keeps going).",
        "operationId": "predict_batch_v1_predict_batch_post",
        "parameters": [
          {
            "in": "query",
            "name": "model_name",
            "required": false,
            "schema": {
              "default": "default",
              "title": "Model Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-api-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchPredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchPredictResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Predict Batch",
        "tags": [
          "predict"
        ]
      }
    },
...
  },
...
}